requirements.txt
altgraph==0.17
appdirs==1.4.4
APScheduler==3.6.3
attrs==20.3.0
cached-property==1.5.2
Cerberus==1.3.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
future==0.18.2
idna==2.10
installer==0.1.1
keyboard==0.13.5
lxml==4.6.3
orderedmultidict==1.0.1
packagebuilder==0.1.0
packaging==20.9
pefile==2019.4.18
pep517==0.10.0
pikepdf==2.11.2
Pillow==8.2.0
pip-shims==0.5.3
plette==0.2.3
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
pynput==1.7.3
pyparsing==2.4.7
PySocks==1.7.1
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
requests==2.25.1
requirementslib==1.5.16
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
stem==1.8.0
toml==0.10.2
tomlkit==0.7.0
tornado==6.1
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
vistir==0.5.2
yaspin==1.5.0
